Rename ambiguous catalog 'utils' package to catutils
This way the package name and searches or types on it will be unique.
Also see: https://blog.golang.org/package-names
