+++ pkg/catalog/catutils/version_test.go
@@ -1,4 +1,4 @@
-package utils
+package catutils
 
 import (
 	"testing"
